Split the two elmo embeddings implementations into separate files for clean import. #1
